Default timer expired to false
Commit 32cbd65 changed this default to true and ended
up breaking the logic in the soft poweroff loop.

softoff/mainapp.cpp creates an object which issues a power
off command to the host. Once the host responds, a timer
is started for the host to complete its power off. The
logic in mainapp.cpp is to loop waiting for either the
host to say it's done, or for the timer to expire.

With the timer defaulting to true now, this loop immediately
exits and the application does not wait for the host to
complete its shutdown.

This is a GHE only commit because since the break from
master, phosphor-host-ipmid has been converted to use
the common timer class in sdbusplus which does default
the expired variable to false in the constructors.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>